aider: Added functionality to reuse the last uploaded image if the checkbox is checked.
# Aider chat conversation:

USER: go on

ASSISTANT: Great! Now that we have the ability to upload an image for the talking head, we need to add the functionality to reuse an already uploaded image.

To do this, we will add a checkbox to the Streamlit interface that allows the user to select whether they want to reuse the last uploaded image. If the checkbox is checked, we will load the image from the file where we saved the last uploaded image. If the checkbox is not checked, we will proceed with the current functionality of uploading a new image.

Here are the changes:

```python
app.py
<<<<<<< HEAD
    # Uploaded image for the talking head
    uploaded_image = st.file_uploader("Upload an image for the talking head", type=["png", "jpg", "jpeg"])
    if uploaded_image is not None:
        with open("uploaded_image.png", "wb") as f:
            f.write(uploaded_image.getbuffer())
        source_url = "uploaded_image.png"
    else:
        source_url = "https://cdn.discordapp.com/attachments/1116787243634397345/1146111608129597450/hypercubefx_face_like_terminator_bb7255e5-efca-489d-bf9e-9aeb750a6bef.png"
=======
    # Option to reuse the last uploaded image
    reuse_last_image = st.checkbox("Reuse last uploaded image")
    if reuse_last_image:
        source_url = "uploaded_image.png"
    else:
        # Uploaded image for the talking head
        uploaded_image = st.file_uploader("Upload an image for the talking head", type=["png", "jpg", "jpeg"])
        if uploaded_image is not None:
            with open("uploaded_image.png", "wb") as f:
                f.write(uploaded_image.getbuffer())
            source_url = "uploaded_image.png"
        else:
            source_url = "https://cdn.discordapp.com/attachments/1116787243634397345/1146111608129597450/hypercubefx_face_like_terminator_bb7255e5-efca-489d-bf9e-9aeb750a6bef.png"
>>>>>>> updated
```

This change adds a checkbox to the Streamlit interface. If the checkbox is checked, the last uploaded image is reused. If the checkbox is not checked, the user can upload a new image. If no image is uploaded, the default image is used.